Moved `tracers.TracerConfig` to its own package to avoid (future) import cycle
This is extracted from https://github.com/ledgerwatch/erigon/pull/10757 to make the overall merge more manageable.
